checking for possibility of VAP existing in cluster before generating binding
